requirements.txt
opencv-python
numpy
requests
ultralytics
huggingface-hub
python-dotenv
python-socketio[client]